Deployment-Server/requirements.txt
flask==3.0.0
anyio==4.0.0
certifi==2023.7.22
charset-normalizer==3.3.0
exceptiongroup==1.1.3
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
notion-client==2.0.0
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
sniffio==1.3.0
urllib3==2.0.7
pytz==2023.3.post1
kafka-python==2.0.2
gunicorn==21.2.0